requirements.txt
dask_image==2023.8.1
filterpy==1.4.5
gphoto2==2.5.0
greenlet==3.0.3
imutils==0.5.4
ipyfilechooser==0.6.0
ipython==8.12.3
ipython==8.14.0
keyboard==0.13.5
lap==0.4.0
matplotlib==3.5.1
moviepy==1.0.3
natsort==8.4.0
numba==0.59.1
numpy==1.24.3
numpy==1.21.5
opencv_python==4.7.0.72
Pillow==9.0.1
Pillow==10.3.0
PyGObject==3.42.1
pymavlink==2.4.39
pypylon==3.0.1
PyQt5==5.15.10
PyQt5_sip==12.9.1
pyqtgraph==0.13.4
PyTurboJPEG==1.7.3
scipy==1.8.0
skimage==0.0
thread==1.1.1
torch==2.2.2
torchsummary==1.5.1
torchvision==0.17.2
